Changed MajorityVote to consider dynamic data
Used a template from canvasjs.com for plotting dynamic data into a bar chart.
The template has a different purpose and thus I have begun making changes to
fit the goals of MajorityVote. I have currently changed the title and what
each bar represents (the name of the dining halls). Currently, the
premade code uses an automatic function that changes the values of the
bars. I need to work on changing the dynamic data to keep track button presses
that represent people clicking on the button to choose where they are going. The
chart also needs to reset during certain periods of the day to indicate
which meal hour it is (breakfast, lunch, dinner).