Made It So You Must Login To Use Groups
Previously, if a user was not logged in, they would still have access to
submit a group name. While this wouldn't have worked, this would cause
an error on the screen and the user would be forced to refresh the page.
So to avoid this messy output, I made it so you have to login or else
no buttons or anything will appear on the screen except a message saying
to login. Thus, no messy error will appear as there is no button for a
non-logged in user to click and they will be informed to login.